small tweaks to simplify deployment
