+++ usermgmt/deploy-local.yml
@@ -24,7 +24,7 @@
     - {src: '{{winsrcdir}}/StationMaint.ps1', dest: '{{windestdir}}/', mode: '754', backup: yes, directory_mode: no }
     - {src: '{{winsrcdir}}/stationMaint2.py', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
     - {src: '{{winsrcdir}}/testkeys.py', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
-    - {src: '{{winsrcdir}}/camTable.py', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
+    - {src: '{{winsrcdir}}/camTableMaintenance.py', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
     - {src: '{{winsrcdir}}/camera.ico', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
     - {src: '{{winsrcdir}}/requirements.txt', dest: '{{windestdir}}/', mode: '644', backup: yes, directory_mode: no }
   - name: update source bucket
