+++ usermgmt/windows/stationMaint2.py
@@ -13,12 +13,102 @@ import json
 import time
 import datetime
 from scp import SCPClient
+from boto3.dynamodb.conditions import Key
+import pandas as pd
 
-from camTable import addRow, getCamUpdateDate, loadLocationDetails
-from camTable import findLocationInfo, dumpCamTable, cameraExists
+
+def addRow(newdata=None, stationid=None, site=None, user=None, email=None, ddb=None, 
+           direction=None, camtype=None, active=None, createdate=None, tblname='camdetails'):
+    '''
+    add a row to the CamTimings table
+    '''
+    if not ddb:
+        ddb = boto3.resource('dynamodb', region_name='eu-west-2')
+    if not newdata:
+        newdata = {'stationid': stationid, 'site': site, 'humanName':user, 'eMail': email, 
+                   'direction': direction, 'camtype': camtype, 'active': active, 'created': createdate,
+                   'oldcode': stationid}
+    table = ddb.Table(tblname)
+    response = table.put_item(Item=newdata)
+    if response['ResponseMetadata']['HTTPStatusCode'] != 200:
+        print(response)
+    return 
+
+
+def loadLocationDetails(table='camdetails', ddb=None, loadall=False):
+    if not ddb:
+        ddb = boto3.resource('dynamodb', region_name='eu-west-2')
+    table = ddb.Table(table)
+    res = table.scan()
+    # strictly, should check for LastEvaluatedKey here, in case there was more than 1MB of data,
+    # however that equates to around 30,000 users which i hope we never have... 
+    values = res.get('Items', [])
+    camdets = pd.DataFrame(values)
+    camdets.sort_values(by=['stationid'], inplace=True)
+    if not loadall:
+        camdets.dropna(inplace=True, subset=['eMail','humanName','site'])
+    camdets['camtype'] = camdets['camtype'].astype(int)
+    camdets['active'] = camdets['active'].astype(int)
+    return camdets
+
+
+def findLocationInfo(srchstring, ddb=None, statdets=None):
+    if statdets is None:
+        statdets = loadLocationDetails(ddb=ddb) 
+        statdets = statdets[statdets.active==1]
+    s1 = statdets[statdets.stationid.str.contains(srchstring)]
+    s2 = statdets[statdets.eMail.str.contains(srchstring)]
+    s3 = statdets[statdets.humanName.str.contains(srchstring)]
+    s4 = statdets[statdets.site.str.contains(srchstring)]
+    srchres = pd.concat([s1, s2, s3, s4])
+    srchres.drop(columns=['oldcode','active','camtype'], inplace=True)
+    return srchres
+
+
+def cameraExists(stationid=None, location=None, direction=None, ddb=None, statdets=None):
+    if statdets is None:
+        statdets = loadLocationDetails(ddb=ddb) 
+        statdets = statdets[statdets.active==1]
+    if stationid:
+        if len(statdets[statdets.stationid == stationid]) > 0:
+            return True
+    if location and direction:
+        s1 = statdets[statdets.site == location]
+        if len(s1) > 0:
+            if len(s1[s1.direction == direction]) > 0:
+                return True
+    return False
+
+
+def dumpCamTable(outdir, statdets=None, ddb=None, exportmindets=False):
+    if statdets is None:
+        statdets = loadLocationDetails(ddb=ddb) 
+        statdets = statdets[statdets.active==1]
+    if exportmindets:
+        statdets = statdets[['stationid', 'eMail']]
+    statdets.to_csv(os.path.join(outdir,'camtable.csv'), index=False)
+
+
+def getCamUpdateDate(camid, ddb=None):
+    if not ddb:
+        ddb = boto3.resource('dynamodb', region_name='eu-west-2')
+    table = ddb.Table('LiveBrightness')
+    resp = table.query(KeyConditionExpression=Key('camid').eq(camid),
+                       IndexName = 'camid-CaptureNight-index',
+                       ScanIndexForward=False,
+                       Limit=1,
+                       Select='SPECIFIC_ATTRIBUTES',
+                       ProjectionExpression='CaptureNight')
+    if len(resp['Items']) > 0:
+        return int(resp['Items'][0]['CaptureNight'])
+    else:
+        return 0
 
 
 class srcResBox(tk.Toplevel):
+    '''
+    A class for the search dialog
+    '''
     def __init__(self, title, message):
         tk.Toplevel.__init__(self)
         self.details_expanded = False
@@ -48,6 +138,9 @@ class srcResBox(tk.Toplevel):
 
 
 class infoDialog(simpledialog.Dialog):
+    '''
+    A class to gather or display info on a camera
+    '''
     def __init__(self, parent, title, location, user, email, sshkey='', id=''):
         self.data = []
         self.data.append(id)
@@ -123,6 +216,9 @@ class infoDialog(simpledialog.Dialog):
 
 
 class statOwnerDialog(simpledialog.Dialog):
+    '''
+    A class to display station ownership data
+    '''
     def __init__(self, parent):
         self.stationdetails = parent.stationdetails
         self.parent = parent
@@ -148,7 +244,9 @@ class statOwnerDialog(simpledialog.Dialog):
 
 
 class CamMaintenance(Frame):
-
+    '''
+    The main camera maintenance window class
+    '''
     def __init__(self, parent):
         self.parent = parent
         Frame.__init__(self, parent)
